ts_packet: remove prototype unsafe implementation, restore old impl
The "new" API was an experiment in reducing allocation and copying, by having
a special backing buffer type that could be unsafely manipulated as a batch.
It sounded promising at the time, but it ended up being cumbersome to use.
We decided to abort that experiment and revisit later on, if/when profiling
supports additional complexity there.

Fixes #29
