+++ ts_overlay_router/src/inbound.rs
@@ -4,7 +4,7 @@ use std::collections::HashMap;
 
 use itertools::Itertools;
 use ts_bart::{RoutingTable, Table};
-use ts_packet::old::PacketMut;
+use ts_packet::PacketMut;
 use ts_transport::OverlayTransportId;
 
 /// An inbound routing action.
