+++ ts_runtime/src/dataplane.rs
@@ -6,7 +6,7 @@ use kameo::{
 };
 use tokio::sync::mpsc;
 use ts_keys::NodePublicKey;
-use ts_packet::old::PacketMut;
+use ts_packet::PacketMut;
 use ts_transport::{OverlayTransportId, UnderlayTransportId};
 
 use crate::{
